repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/url_shortener/18/Reflexion/url_shortener.py
import random
import string
from flask import Flask, request, redirect

app = Flask(__name__)

# Mock database
DATABASE = {}

@app.route('/shorten_url', methods=['POST'])
def shorten_url():
	url = request.json['url']
	short_url = ''.join(random.choices(string.ascii_uppercase + string.digits, k=5))
	DATABASE[short_url] = url
	return {'short_url': short_url}

@app.route('/<short_url>', methods=['GET'])
def redirect_url(short_url):
	url = DATABASE.get(short_url)
	if url:
		return redirect(url)
	else:
		return {'error': 'URL not found'}, 404
